Require directories in the image archive to be managed by the helioviewer group (#157)
This change requires that the server running Helioviewer has a group called 'helioviewer'

This group gets write permissions to the helioviewer image archive specified in other configuration files. This allows multiple users on the system to be able to manage helioviewer files without needing superuser privileges to make repairs to the image archive. This also allows `downloader.py` to be used by multiple users without causing any permission conflicts in the image archive.