Shrink time range for vso_sunpy_download test
This removes the upstream dependency on this test.
The test generates a python script for downloading files and executes it.
If the files can't be downloaded from the upstream source, the test will fail.
This tightens the time range so there are no files to be downloaded.
It does not completely remove the upstream dependency, but it
reduces it.