Download support for GOES-19 CCOR-1 (#437)
* ingest support for GOES-19 CCOR-1
also includes partial support for SWFO-L1 CCOR.

* Rename GOES-R/GOES-19 to GOES

* set CCOR2 data server to TBD

* Fix punch in db.py

* Shrink time range for vso_sunpy_download test

This removes the upstream dependency on this test.
The test generates a python script for downloading files and executes it.
If the files can't be downloaded from the upstream source, the test will fail.
This tightens the time range so there are no files to be downloaded.
It does not completely remove the upstream dependency, but it
reduces it.

* commit test file

* Mock requests and responses for sunpy_download test

* Add responses to test requirements

* Isolate CUNIT patch for GOES

* fix patch and update comment